+++ src/page/TableListPage/TableListPage.tsx
@@ -1,5 +1,5 @@
 import DefaultLayout from '../../layout/defaultLayout/DefaultLayout';
-import Table from '../LoginPage/components/Table';
+import Table from './components/Table';
 
 const data = [
   { name: '테이블1', type: '의회식 토론', time: 30 },
@@ -22,7 +22,7 @@ export default function TableListPage() {
         <DefaultLayout.Header.Left>테이블 목록화면</DefaultLayout.Header.Left>
       </DefaultLayout.Header>
       <div className="flex h-screen flex-col border-2 border-red-200 px-4 py-6">
-        <main className="grid grid-cols-3 justify-items-center gap-4">
+        <main className="grid grid-cols-3 justify-items-center gap-6">
           {data.map((table: TableProps, idx: number) => (
             <Table
               key={idx}
