+++ src/page/TableListPage/components/AddTable.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function AddTable() {
+  return <div>AddTable</div>;
+}
